Split inference and chat
